+++ dataservices/models.py
@@ -331,22 +331,13 @@ class UKTradeInServiceByCountry(TimeStampedModel):
     value = models.DecimalField(max_digits=20, decimal_places=2, null=True, blank=True)
 
 
-class UKTotalTradeByCountry(models.Model):
-    DIRECTIONS = [
-        ('IMPORT', 'Import'),
-        ('EXPORT', 'Export'),
-    ]
-    TYPES = [
-        ('GOODS', 'Goods'),
-        ('SERVICES', 'Services'),
-    ]
+class UKMarketTrends(models.Model):
     country = models.ForeignKey(
         'dataservices.Country', verbose_name=_('Countries'), on_delete=models.SET_NULL, null=True
     )
-    year = models.IntegerField(null=True, blank=True)
-    direction = models.CharField(max_length=15, choices=DIRECTIONS)
-    type = models.CharField(max_length=15, choices=TYPES)
-    value = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+    year = models.PositiveSmallIntegerField(null=True, blank=True)
+    imports = models.PositiveIntegerField(null=True, blank=True)
+    exports = models.PositiveIntegerField(null=True, blank=True)
 
     class Meta:
-        verbose_name = "UK total trade by country"
+        verbose_name = "UK market trends by country"